feat(cva)!: split into a lean cva/core engine and a clsx-wired preset
Restructure around a new `cva/core` entry point whose `defineConfig`
requires a `cx` concatenator and never interprets class values: core
assembles the authored pieces (composed outputs, base, matched variant
and compound values, class/className) and passes them to the
concatenator verbatim, one argument each, with core-created arrays
spread flat. The concatenator owns the grammar entirely, so twMerge,
clsx/lite, cnfast's cn, and clsx all work within their own input
contracts — including clsx/lite for string-authored components.

The main `cva` package becomes the batteries preset: cva/cx/compose
wired up with real clsx for class-variance-authority parity ("cx is an
alias of clsx" again), plus a deprecated defineConfig re-export (cx
optional, defaulting to clsx) pointing migrators at cva/core.

Typing is handled by a new augmentable CVARegistry interface: the cx
option and the whole authoring surface default to ClassValue, and one
`declare module "cva/core"` line narrows both to a concatenator's own
input type (e.g. tailwind-merge's ClassNameValue), making bare
`cx: twMerge` typecheck while turning object-syntax authoring into a
compile error project-wide. Registry narrowing is verified in an
isolated compile (type-tests/ + check:registry) since module
augmentation is program-wide and would poison the main suite.

Adds a real-package concatenator matrix (concatenators.test.ts) with
tailwind-merge and cnfast as devDependencies alongside clsx/clsx/lite,
pinning type assignability and runtime behavior (conflict resolution,
grammar support, lite's string-only dropping) for each.

Rebuilt on top of the tsdown migration (0c6ca91), which carries the
multi-entry packaging natively: src/core.ts is a second tsdown entry,
the exports/publishConfig.exports maps are build-generated, rolldown
emits the cross-entry import with correct per-format extensions, and
./core's node10 fallback is a hand-maintained publishConfig
typesVersions block following the stable package's ./types pattern.
All build gates (attw strict, publint, unused) pass for both entries;
dist/index.cjs 1.25kB and dist/core.cjs 986B, within budget.
