+++ AGENTS.md
@@ -150,7 +150,9 @@ Agent-specific notes:
 
 Durable, hard-won lessons that don't fit a section above. See [Keeping this guide current](#keeping-this-guide-current-self-improving) for what belongs here and how to write it. Newest first; prune anything that's become wrong or obsolete.
 
-- Both published packages (`packages/cva`, `packages/class-variance-authority`) build with [`tsdown`](https://tsdown.dev) — the full pipeline (publish transform, config feature reference, what each build-time gate checks) is documented in [CONTRIBUTING.md's Build & publish section](./CONTRIBUTING.md#build--publish-packages); read it before touching either build. The agent-critical gotchas: each `package.json`'s `exports`/`publishConfig.exports` blocks are **regenerated on every build** — never hand-edit them, change that package's `tsdown.config.mts` (or the shared `.config/tsdown.base.mts`) instead (exception: `class-variance-authority`'s hand-maintained `publishConfig.typesVersions`, which tsdown preserves but doesn't generate); a red build from the attw/publint/unused gates means the _publish shape_ broke, not the source (tsdown's pack step applies `publishConfig` — verified by breaking `dts` and watching the gates fail); and always `pnpm pack`, never `npm pack`, since only pnpm applies the `publishConfig` exports rewrite.
+- TypeScript module augmentation is program-wide, so `CVARegistry`-narrowing assertions can never live in `packages/cva`'s main test suite — one `declare module "cva/core"` in any test file rewrites the authoring types for every other test in the same compilation. Registry type-tests live in the isolated `packages/cva/type-tests/` compile instead (`.config/tsconfig.type-tests.json`, run via `check:registry`), which maps the package's own name to source via `paths` so fixtures read exactly like consumer code. Note TS 6 deprecates `baseUrl`: `paths` entries resolve relative to the tsconfig file itself.
+- `clsx/lite`'s published types lie: the `./lite` export reuses the full build's `clsx.d.mts` (claiming `ClassValue[]` inputs), but its runtime only appends string arguments and silently drops arrays/objects/numbers. Never rely on lite's types for safety; it works as a `cva/core` concatenator for string-authored components only because core flat-spreads its own arrays and passes authored values through verbatim (pinned in `packages/cva/src/concatenators.test.ts`).
+- Both published packages (`packages/cva`, `packages/class-variance-authority`) build with [`tsdown`](https://tsdown.dev) — the full pipeline (publish transform, config feature reference, what each build-time gate checks) is documented in [CONTRIBUTING.md's Build & publish section](./CONTRIBUTING.md#build--publish-packages); read it before touching either build. The agent-critical gotchas: each `package.json`'s `exports`/`publishConfig.exports` blocks are **regenerated on every build** — never hand-edit them, change that package's `tsdown.config.mts` (or the shared `.config/tsdown.base.mts`) instead (exception: the hand-maintained `publishConfig.typesVersions` blocks — `class-variance-authority`'s `./types` and `cva`'s `./core` node10 fallbacks — which tsdown preserves but doesn't generate); a red build from the attw/publint/unused gates means the _publish shape_ broke, not the source (tsdown's pack step applies `publishConfig` — verified by breaking `dts` and watching the gates fail); and always `pnpm pack`, never `npm pack`, since only pnpm applies the `publishConfig` exports rewrite.
 - `packages/cva`'s type-export rule is in [`CONTRIBUTING.md`'s Style Guide](./CONTRIBUTING.md#style-guide) — `packages/cva/src/index.test.ts`'s `describe("exported types", ...)` and `describe("CVAVariantShape", ...)` blocks pin the current set of portability exports (`CVAComponent`/`CVAComponentShape`/`CVAVariantShape`) via the `CVA.` namespace import, so losing one immediately fails `pnpm --filter cva check:tsc` — but that guard only catches losing an _existing_ export, not a _new_ internal type leaking into a public signature. To check for the latter after reshaping `cva`'s public API, verify manually against the packed artifact: `pnpm --filter cva build`, then `pnpm --filter cva pack --pack-destination <tmp>` (use `pnpm pack`, never `npm pack` — only pnpm applies the package's `publishConfig` exports rewrite), extract with `tar -xzf <tmp>/cva-*.tgz --strip-components=1 -C <tmp>/node_modules/cva` (the tarball has a top-level `package/` directory that needs stripping, or `cva` won't resolve), write a small consumer file re-exporting a component / `VariantProps<typeof x>` / `getSchema` result, and compile it with `tsc --strict --declaration --emitDeclarationOnly --module nodenext --moduleResolution nodenext` (the module flags matter — without them `tsc` resolves via the `types` field and never traverses the rewritten `exports` map that consumers actually hit).
 - `pr-comment.yml` (`workflow_run`) can't be exercised from a PR — GitHub always runs a `workflow_run` workflow's file as it exists on the default branch, so changes to it are invisible to the triggering PR's own run. The risky parts — [`test/bench/scripts/compare.ts`](./test/bench/scripts/compare.ts)'s validation/rendering, [`.github/scripts/process-pr-comment.mjs`](./.github/scripts/process-pr-comment.mjs)'s artifact/PR binding, and [`.github/scripts/pr-comment.mjs`](./.github/scripts/pr-comment.mjs)'s section-splicing — are covered by their own test files instead; after merging a change to the workflow glue, verify end-to-end with a scratch PR from a fork (benchmark comment appears, stale-run skip works, no duplicate sticky comments).
 - The benchmark job must not write an Actions summary for PRs: PR code runs earlier in the same runner and can influence later steps. The separate `workflow_run` comment job uses a fresh runner and trusted default-branch renderer; keep the job summary limited to trusted pushes to `main`.
